requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
pillow==10.4.0
pyTelegramBotAPI~=4.22.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3

pip~=24.2
amqp~=5.2.0
vine~=5.1.0
kombu~=5.4.2
billiard~=4.2.1
celery~=5.5.0b3
click~=8.1.7
python-dateutil~=2.9.0.post0
click-repl~=0.3.0
click-plugins~=1.1.1
wcwidth~=0.2.13
six~=1.16.0
prompt_toolkit~=3.0.47
h11~=0.14.0
sniffio~=1.3.1
httpcore~=1.0.5
anyio~=4.6.0
python-telegram-bot~=21.6
httpx~=0.27.2